projekt/src/aplikacja/requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
pbr==5.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
stevedore==1.31.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
